# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ركين - منصة الوساطة الإلكترونية API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
  phone: string;
  countryCode: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  broker: 'broker',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  phone: string;
  countryCode: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface Broker {
  id: number;
  userId: number;
  username: string;
  email: string;
  phone: string;
  transactionCount: number;
  successRate: number;
  avgTime: string;
  rating: number;
  createdAt: string;
}

export interface BrokerInput {
  userId: number;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  active: 'active',
  paid: 'paid',
  completed: 'completed',
  closed: 'closed',
  cancelled: 'cancelled',
} as const;

export interface Transaction {
  id: number;
  title: string;
  description: string;
  amount: number;
  type: string;
  status: TransactionStatus;
  buyerId: number;
  sellerId: number;
  brokerId: number;
  /** @nullable */
  buyerName?: string | null;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  brokerName?: string | null;
  createdAt: string;
}

export interface TransactionInput {
  title: string;
  description: string;
  amount: number;
  type: string;
  brokerId: number;
  sellerEmail: string;
}

export type MessageMessageType = typeof MessageMessageType[keyof typeof MessageMessageType];


export const MessageMessageType = {
  text: 'text',
  system: 'system',
  payment: 'payment',
} as const;

export interface Message {
  id: number;
  transactionId: number;
  senderId: number;
  senderName: string;
  /** @nullable */
  senderRole?: string | null;
  content: string;
  messageType: MessageMessageType;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface BrokerTransferInput {
  email: string;
  amount: number;
}

export interface BrokerTransferResult {
  recipientName: string;
  newBalance: number;
}

export interface WalletBalance {
  balance: number;
}

export interface WithdrawInput {
  amount: number;
}

export interface PlatformStats {
  userCount: number;
  transactionCount: number;
  successRate: number;
  brokerCount: number;
  totalAmount: number;
}