# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ركين - منصة الوساطة الإلكترونية API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /brokers:
    get:
      operationId: listBrokers
      tags: [brokers]
      summary: List all brokers
      responses:
        "200":
          description: List of brokers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Broker"
    post:
      operationId: createBroker
      tags: [brokers]
      summary: Add a broker (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrokerInput"
      responses:
        "201":
          description: Broker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broker"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /brokers/{id}:
    delete:
      operationId: deleteBroker
      tags: [brokers]
      summary: Remove a broker (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Broker removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /transactions/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages for a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    RegisterInput:
      type: object
      required: [username, email, password, phone, countryCode]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 8
        phone:
          type: string
        countryCode:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, username, email, phone, countryCode, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        countryCode:
          type: string
        role:
          type: string
          enum: [user, broker, admin]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    Broker:
      type: object
      required: [id, userId, username, email, phone, transactionCount, successRate, avgTime, rating, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        transactionCount:
          type: integer
        successRate:
          type: number
        avgTime:
          type: string
        rating:
          type: number
        createdAt:
          type: string
    BrokerInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
...
    Message:
      type: object
      required: [id, transactionId, senderId, senderName, content, messageType, createdAt]
      properties:
        id:
          type: integer
        transactionId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderRole:
          type: ["string", "null"]
        content:
          type: string
        messageType:
          type: string
          enum: [text, system, payment]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
...