# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ركين - منصة الوساطة الإلكترونية API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
  phone: string;
  countryCode: string;
}